Added more driver functionality.
Added device file get_serial. This will query the keyboard, firefly and mamba for the devices serial number.
The file is meant to be read from and not written to. Will go through the drivers and set permissions appropiatly
to restrict either reading or writing.

Also added a little bit to makefile