Added Razer Stealth serial
If the Razer Stealth attempts to get a serial it will return 15 X's, this wont clash with other serials as they are of the format XX999999999999, where X is an uppercase letter and 9 represents any digit.
Moved `row_length` line to top of the fuction to remove a compiler warning.