Have commented out device busy notifications
Did some tests and it seems the device still respects what it was sent so I dont think its an error so much as an informational message.
Calculated some timings and whilst we do spam the kb more than synapse even if I do slow it down to synapse speed it stil complains but also works so
am ignoring it for now. Seems ok.