Rework stop_daemon to rely on the pidfile
Instead of relying on grep and pgrep, rely on the pidfile data. On the
assumption that if we start daemons with different users, they probably
want to shoot themselves in the foot...

Relying on the pid file also requires us to check the name on the process to
avoid shooting down a process that recycled the pid. This race existed with
the current code but was less likely to trigger.

The delay is now up to 3s, instead of always 3s and the delay after stop() is
moved to the respawn() code, it doesn't apply otherwise anyway.