driver: invert the blade logo state in the driver
Apparently some (?) of the blade devices have the logo led state inverted.
This was previously worked around by having custom dbus methods that handle
the inversion but this is simply stuff that should be handled in the driver so
that userspace always sees the right behavior and doesn't have to care about
hardware-specifics.

The list of devices was taken from the dbus daemon of those that have a
workaround.

This is untested, I don't have a device

Fixes #389