Openrazer daemon cleanup patches (#385)
* daemon: s/logging_level/log_level/

if we use 'log_dir' we should use 'log_level' for consistency

* daemon: survive failure to open screen saver

If we want to run the daemon on the system bus, we don't have an X connection.

* daemon: move existing process check to a helper function

* daemon: move creation of logger to helper function

* daemon: move logger initialization up

So we have the logger available everywhere

* daemon: move check for plugdev group to helper function

* daemon: always assume that root is in plugdev

Because, well, it's the root user

* daemon: if the 'plugdev' group lookup failed, the user isn't in it

If we cannot look up the group, the user cannot be a member of the group

* daemon: move check for plugdev group up

No point starting a dbus daemon if we're about to quit because of our group
membership

* daemon: move udev monitor creation into a helper function

* daemon: move screensaver init to helper function

* daemon: use 'BUS_NAME', not 'BUS_PATH'

Let's use the correct technical terms here

* daemon: make the dbus method initialization easier to parse

* daemon: remove a couple of needless checks

The defaults guaranteed these were always set anyway, we can just set those
normally in the dictionary

* daemon: only add files ending in .py to the module import list

vim creates .foo.py.swp files while a file is open which causes the daemon to
fail with confusing error messages:
ModuleNotFoundError: No module named 'openrazer_daemon.hardware.'

* daemon: remove unnecessary os.path.exists() check

os.path.isdir() won't ever return true if it doesn't exist

* daemon: if the run_dir exists but is not a directory, fail early

* daemon: create the run_dir if it doesn't exist

If a run_dir is defined, but it doesn't exist yet, just create it. We already
fail for a run_dir that exists and is a file, so we can ignore that bit now.

* daemon: don't check for existence before os.makedirs

We're making the directories anyway, so let python handle the errors. Bonus
point: we now handle the case where the target path exists but is a file

* daemon: add the __name__ check for __main__

No real changes, just means it's really only run when being run as script

* daemon: defer making directories until we need them

Move the run_dir and log_dir into the daemon proper, we shouldn't create them
unless we actually start the daemon. Same goes for the example config file, it
should not be created until we're sure we want to start the daemon.

* daemon: fail if the given config file does not exist

If the user specifies a config file it must exist.