daemon: if we're running on foreground mode, supply our own logger
daemonize initializes a logger during start(). Any exception in the 'action'
argument will be logged to that logger - which defaults to the syslog.

Inside run_daemon() we set up our own logger for the daemon itself and catch
any exceptions by logging to that. However, if an exception occurs during the
daemon's __init__(), that exception was only caught by daemonize and thus
logged to syslog only.

Avoid this by using another logger in case we're using -F

The proper fix is likely to have a classmethod that returns the logger for the
daemon and re-use that for daemonize too. But that requires moving the log
directory creation etc. into the startup script.

Fixes #431