Work around race condition when hotplugging
The problem is that udev throws 'add' events in a random order but the
current code needs the interface with our sysfs entries first, and then
the other interfaces which get added as 'additional interfaces'. Without
this, in most cases the interfaces get added in the wrong order,
resulting in ripple, game mode, macro functionality etc not working

Now the to-be-added devices get batched up and sorted and only then
really added.

A downside of this commit is that adding new devices is getting delayed by at least 2 seconds (maybe this time can be lowered).

Test plan:
1. Restart daemon while keyboard is plugged in
(optionally verify that ripple and game mode works)
2. Unplug keyboard
3. Replug keyboard
4. Notice that ripple and game mode doesn't work without this commit and works with this commit.

Fixes #390