Fix effect sync between keyboard and mice
Prior to this change the only effects that synced between the mouse and keyboard when effect sync was enabled and I selected the effect on the keyboard were:

static
spectrum

Other effects were failing to be applied to the mouse's scroll, logo, and backlight. (Lancehead TE LEDs aren't supported yet)

This change fixes the following effects to be applied by the mouse's scroll, logo, and backlight (if a mouse has one) when the effect is selected on the keyboard and effect sync is enabled:

wave
reactive
none
breath random
breath dual
breat single

I tried to preserve the logic for pulse effect devices, but do not have a pulse effect device to test with.