Fix brightness sync between keyboard and mouse
Prior to this change:

If the user had a mouse and keyboard with sync devices enabled, then modified the brightness on the keyboard, the mouse LEDs would not sync the effect setBrightness, scroll, logo, backlight properly.

With this change:

When the brightness on the keyboard updates, the scroll, logo, and backlight LEDs on the mouse properly update when sync devices is enabled.