Add persistent storage of effects in daemon
This patch adds several D-Bus methods for accessing the current state of
the devices (plus pylib API for the new functionality):
* get*Effect
* get*EffectColors
* get*EffectSpeed
* get*WaveDir
* restoreLastEffect

The star means e.g. getEffect, getLogoEffect, getScrollEffect, etc.

The last one can be used to reset to the previous effect after setting a
custom effect (which won't be persisted).

Co-authored-by: Luke Horwell <code@horwell.me>
Co-authored-by: Luca Weiss <luca@z3ntu.xyz>