Translate Profile and Clutch buttons to F18 and F19
There seems to be parallel construction between the remapping in this
driver and the M1-M5 key handling in the razerkbd driver. Since razerkbd
assigns F17 to input code 0x24, I didn't want to use F17 to mean some
other code in the mouse driver. Hence, F17 is skipped and the Basilisk's
new buttons are mapped starting at F18.