create_fake_device.py: remove --create-only option
This option creates problem with the event files.

Reason is that the created event files are fifo pipes which - if not
opened on this side - will block on the other side (the daemon) and hang
forever. So we need to keep these pipes open to not make the daemon
misbehave.