daemon: fix misconfigured matrix capabilities
Devices were throwing an exception that was:
- "embedded null byte"
- or "setCustom is not a valid method of interface"

This was due to devices missing a sysfs file in the driver. Or rather,
they had one (matrix_effect_custom), but was incorrectly specified here.

For devices that are missing a sysfs file, it's possible custom frames
are unsupported or was simply left out in the driver when the device was
added. Users of these devices should try amending their device in the
driver, as we don't have the hardware to check.

Fixes #1252
Addresses polychromatic/polychromatic#266