Remove Gdk dependency
Removing the dependency prevents a crash when the "pylib" is used in an
application/script that uses PyQt5/PySide2 under specific conditions.

The "pylib" library indirectly inherited Gdk as it imports the keyboard
and misc.macro modules from openrazer_daemon. In practice, Gdk isn't
used at all, so it is safe to remove. This looks to be legacy code as
gdk_colour_to_rgb() was always returning a tuple.

Fixes polychromatic/polychromatic#317
Fixes polychromatic/polychromatic#319