pylib: Fix UnknownMethodException for charging zones
When attempting to use the Python library for the 'fast_charging' or
'fully_charged' zones, an UnknownMethodException was thrown due to an
extra underscore.

This caused "setFastChargingWave" to become "setFast_ChargingWave" which
is incorrect. The method path is correct to use an underscore:
razer.device.lighting.fast_charging

The other zones are unaffected as underscores are not used for the DBUS
method name itself (e.g. setLogoStatic, setScrollStatic, etc)