daemon: Pass Y=0 for DPI X only devices
The DPI for X-only mice could not be set via the pylib due to a minimum
check of zero for the Y value. The daemon expected a value of -1 for
such devices.

However, -1 cannot be submitted via D-Bus because it's an unsigned
integer (cannot be negative). To address this, this commit changes the
value to 0.

Up to now, frontends using the Python library could not pass -1, and
passing 0 did not work either. This fixes setting DPI for those devices.