driver: change device_mode read
Previously we wrote bytes to device_mode and read an ASCII string. This
doesn't make much sense, so fix it. We want to keep daemon API
compatibility though so adjust the daemon to still emit the same ascii
string.

This also makes it possible to use device_mode with the fake driver as
we can just read back the bytes we wrote to it.

In OpenRazer this shouldn't have a functional change but if any 3rd
party application relies on reading ascii from device_mode, that will
break.