daemon: stop silently constraining DPI value to max
If something is trying to set DPI to more than the maximum, that's a bug
and the user should know by throwing an exception. Also move the check
further up in the function so at least it triggers before saving any
values to persistence.

Also we can assume DPI_MAX always exists, if not that's a bug as well.