daemon: notify for new effect when calling setCustom/setKeyRow
The RippleManager needs to know when to stop itself and since this
happens via the notify mechanism, we do need to tell the ripple to stop
when someone wants to set a custom effect.

However we need to make sure to not stop when ripple is itself using the
APIs, to make a 'private' version of setKeyRow and setCustom that
doesn't have the notify code, and only add the notify to the public
DBus-accessible function.

Fixes #1714