Remove "None" effect for Basilisk V3
Setting the "None" effect causes this mouse to enter a buggy state.
If the computer was rebooted while the "None" is set, the mouse won't be
detected again until it is physically replugged.

Some users found that a firmware update solved it, while others reported
a factory reset (on Synapse) works. Ultimately, the firmware is stable
when none of the "None" effects are set.

Better to prevent this situation from happening by removing the effect.

Fixes #2049
Addresses #1767, #1844