daemon: Relax serial number check a bit
While many serial numbers have two uppercase letters and then a mix of
numbers and letters, there seems to be also valid serial numbers which
start with a number. Relax the check to allow any mix of numbers and
uppercase letters. Let's see if that's good enough or we need to tweak
it further.

Closes: #2403
Fixes: a32dc5ff ("daemon: Be more strict in accepting serial numbers")