driver: Stop copying any custom frame data when parameters are invalid
While the initial idea of this check was to sanitize any a row_length
value which is too high. But in reality we should just essentially error
out (which due to the function signature we can't do properly), and stop
trying to memcpy any data.

Issue: https://github.com/openrazer/openrazer/issues/2433