driver: Update set_bit code to use DECLARE_BITMAP
There seems to be a exc_alignment_check happening and killing the kernel
when we cast our u8[20] to unsigned long* and setting the bit, probably
because 20 bytes don't fit well into longs, which are 8 byte long each.

Just use the regular DECLARE_BITMAP() macro which should resolve the
issue. Adjust the code around to match the new size.

Issue: #2439