razeraccessory: Remove 'defaults' for transaction ID
To prepare getting the transaction IDs a bit more unified, start
by replacing the default case by explicit case statements for the
devices that use them.

For now, it makes it explicit that a code path needs to be chosen for
new devices, and in the future we can use this data to investigate how
to unify this and make this simpler.

This change was done in a script-supported way by slightly modifying
generate_fake_driver.sh to output all devices that create a given sysfs
file, then finding the missing ones from the existing case statements
using essentially grep.

No functional change intended.