pylib: Raise NotImplementedError more consistently
Avoid masking issues in the caller, calling methods on devices that
don't support it. Raise an exception if such a situation occurs to let
the caller know. This does change the API for users, but should
hopefully not cause any issues, since callers should already check for
feature availability before calling.